+++ src/mctpd.c
@@ -3388,13 +3388,26 @@ static int del_interface(struct link *link)
 	if (ctx->verbose) {
 		fprintf(stderr, "Deleting interface #%d\n", ifindex);
 	}
-	for (size_t i = 0; i < ctx->num_peers; i++) {
+	for (size_t i = 0; i < ctx->num_peers;) {
 		struct peer *p = ctx->peers[i];
+
 		if (p->state == REMOTE && p->phys.ifindex == ifindex) {
-			// Linux removes routes to deleted links, so no need to request removal.
+			int rc;
+
+			// Linux removes routes to deleted links, so no need
+			// to request removal.
 			p->have_neigh = false;
 			p->have_route = false;
-			remove_peer(p);
+			rc = remove_peer(p);
+			if (rc) {
+				bug_warn("Error removing peer on interface "
+					 "deletion, inconsistent state");
+				break;
+			}
+		} else {
+			// Removal will shift indices down, only increment
+			// while skipping.
+			i++;
 		}
 	}